+++ .coveragerc
@@ -1,0 +1,29 @@
+# .coveragerc to control coverage.py
+[run]
+branch = True
+omit =
+    */examples/*
+
+[report]
+# Regexes for lines to exclude from consideration
+exclude_lines =
+    # Have to re-enable the standard pragma
+    pragma: no cover
+
+    # Don't complain about missing debug-only code:
+    def __repr__
+    if self\.debug
+
+    # Don't complain if tests don't hit defensive assertion code:
+    raise AssertionError
+    raise NotImplementedError
+
+    # Don't complain if non-runnable code isn't run:
+    if 0:
+    if __name__ == .__main__.:
+    pass
+
+ignore_errors = True
+
+[html]
+directory = coverage_html_report
