+++ asyncworker/__init__.py
@@ -1,7 +1,7 @@
 from asyncworker.base import BaseApp
 from asyncworker.consumer import Consumer
 from asyncworker.options import Options, Defaultvalues, Events
-from asyncworker.signal_handlers.rabbitmq import RabbitMQ
+from asyncworker.signals.handlers.rabbitmq import RabbitMQ
 
 
 class App(BaseApp):
