+++ asyncworker/signals/handlers/rabbitmq.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING
 
 from asyncworker.consumer import Consumer
-from asyncworker.signal_handlers.base import SignalHandler
+from asyncworker.signals.handlers.base import SignalHandler
 
 if TYPE_CHECKING:  # pragma: no cover
     from asyncworker import App
